AjintK/requirements.txt
# Claude Agent Framework Requirements with Multi-Provider Support

# Core dependencies - LLM Providers
anthropic>=0.40.0
google-generativeai>=0.8.0
mistralai>=1.0.0
requests>=2.31.0
httpx>=0.27.0

# Configuration
pyyaml>=6.0

# Optional: Enhanced CLI features
click>=8.0.0
rich>=13.0.0

# Optional: Advanced logging
python-json-logger>=2.0.0

# Retry mechanism
tenacity>=8.2.0
